Moved fetchPDF from App.js to Api.js
