+++ react/src/components/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import './App.css'
 import CVEditor from './CVEditor'
 import SearchField from './SearchField'
-import { saveCV, loadCV } from './Api'
+import { saveCV, loadCV, fetchPDF } from './Api'
 import Preview from './Preview'
 
 
@@ -44,20 +44,8 @@ class App extends Component {
     )
   }
 
-
-  fetchPDF() {
-    fetch(`api/users/${this.state.uid}/pdf`)
-      .then(res => res.blob())
-      .then((blob) => {
-        const file = new File([blob], `${this.state.uid}.pdf`, { type: 'application/pdf' })
-        const a = document.createElement('a')
-        a.href = URL.createObjectURL(file)
-        a.download = `${this.state.uid}.pdf`
-        document.body.appendChild(a)
-        a.click()
-        document.body.removeChild(a)
-      })
-      .catch(err => console.log(err))
+  exportAsPDFClicked() {
+    fetchPDF(this.state.uid)
   }
 
   render() {
@@ -73,7 +61,7 @@ class App extends Component {
           saveStatus={this.state.saveStatus}
           updateText={text => this.updateText(text)}
           saveCV={() => this.saveCV()}
-          fetchPDF={() => this.fetchPDF()}
+          fetchPDF={() => this.exportAsPDFClicked()}
         />
         <Preview
           text={this.state.text}
